Move sensitivity-vs-role helpers to MCP/access.py + tidy is_expired import
Three review fixes from PR #77:

- ALLOWED_SENSITIVITIES, require_project_membership, and
  require_can_write_at_sensitivity all moved out of MCP/servers/people.py
  into MCP/access.py — none of them are person-specific, and other MCP
  write tools (notes, reports, observations, …) need the same matrix.
  ALLOWED_SENSITIVITIES is now derived from access_control.SensitivityLevel
  rather than a hand-listed frozenset, so adding a sensitivity tier in
  one place stays in sync with the matrix.

- Hoisted `from memory.api.auth import is_expired` to module scope in
  MCP/access.py — was inline inside is_session_expired. auth.py doesn't
  import from memory.api.MCP, so there's no circular-import reason for
  the indirection.

- Test imports updated (memory.api.MCP.servers.people →
  memory.api.MCP.access) per the no-backcompat-shims convention.
